schemas: add RDMParentSchema
